+++ packages/auth-api-token/graphql/Mutation/generateAPIKey.ts
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/no-namespace */
 import { signJwt } from '@zemble/auth/utils/signJwt'
 import { GraphQLError } from 'graphql'
 
@@ -8,7 +7,7 @@ import type { MutationResolvers } from '../schema.generated'
 
 const { API_KEY_SECRET } = plugin.config
 
-const generateAPIKey: MutationResolvers['generateAPIKey'] = async (_: unknown, { expiresInSeconds, apiKeySecret }, { decodedToken }) => {
+export const generateAPIKey: NonNullable<MutationResolvers['generateAPIKey']> = async (_: unknown, { expiresInSeconds, apiKeySecret }, { decodedToken }) => {
   if (apiKeySecret !== API_KEY_SECRET) {
     throw new GraphQLError('Invalid apiKeySecret')
   }