+++ packages/bull/graphql/Mutation/drainQueue.ts
@@ -4,7 +4,7 @@ import { getQueueByName } from '../../utils/setupQueues'
 
 import type { MutationResolvers } from '../schema.generated'
 
-const drainQueue: MutationResolvers['drainQueue'] = async (_, { queue }) => {
+export const drainQueue: NonNullable<MutationResolvers['drainQueue']> = async (_, { queue }) => {
   const q = getQueueByName(queue)
 
   if (!q) {