+++ packages/bull/graphql/Mutation/removeJob.ts
@@ -4,7 +4,7 @@ import { getQueueByName } from '../../utils/setupQueues'
 
 import type { MutationResolvers } from '../schema.generated'
 
-const removeJob: MutationResolvers['removeJob'] = async (_, { queue, jobId }) => {
+export const removeJob: NonNullable<MutationResolvers['removeJob']> = async (_, { queue, jobId }) => {
   const q = getQueueByName(queue)
 
   if (!q) {