+++ packages/cms/graphql/Query/getEntityByNameSingular.ts
@@ -2,7 +2,7 @@ import { readEntities } from '../../utils/fs'
 
 import type { QueryResolvers } from '../schema.generated'
 
-const getEntityByNameSingular: QueryResolvers['getEntityByNameSingular'] = async (_, { name }) => {
+export const getEntityByNameSingular: NonNullable<QueryResolvers['getEntityByNameSingular']> = async (_, { name }) => {
   const entities = await readEntities()
   const result = entities.find((entity) => entity.nameSingular === name)