+++ packages/auth-anonymous/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @zemble/auth-anonymous
 
+## 0.0.40
+
+### Patch Changes
+
+- 80fe6c9: Improve authorization model
+- Updated dependencies [80fe6c9]
+  - @zemble/auth@0.0.42
+
 ## 0.0.39
 
 ### Patch Changes