+++ packages/create-zemble-app/CHANGELOG.md
@@ -1,5 +1,14 @@
 # create-zemble-app
 
+## 0.1.0
+
+### Minor Changes
+
+- 1928354: Move in eslint package, rename graphql-codegen to codegen, other improvements.
+  ```
+
+  ```
+
 ## 0.0.6
 
 ### Patch Changes